+++ src/tag/parser/compile/hud_interface.cpp
@@ -30,6 +30,13 @@
         } \
     } \
 }
+        
+#define CHECK_HIGH_RES_SCALING(flags, name) \
+if((flags & HEK::HUDInterfaceScalingFlagsFlag::HUD_INTERFACE_SCALING_FLAGS_FLAG_USE_HIGH_RES_SCALE) && workload.get_build_parameters()->details.build_cache_file_engine == HEK::CacheFileEngine::CACHE_FILE_XBOX) { \
+    std::string name_copy = name; \
+    name_copy[0] = std::toupper(name_copy[0]); \
+    REPORT_ERROR_PRINTF(workload, ERROR_TYPE_ERROR, tag_index, "%s has high resolution scaling enabled, but this does not exist on the target engine", name_copy.c_str()); \
+}
 
 namespace Invader::Parser {
     void WeaponHUDInterface::pre_compile(BuildWorkload &, std::size_t, std::size_t, std::size_t) {
@@ -137,24 +144,26 @@ namespace Invader::Parser {
 
     void WeaponHUDInterfaceMeter::post_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t struct_offset) {
         // Make sure it's valid
-        if(this->sequence_index != NULL_INDEX && (!workload.disable_recursion || !workload.disable_error_checking)) {
-            CHECK_INTERFACE_BITMAP_SEQ(this->meter_bitmap, this->sequence_index, (std::string("meter #") + std::to_string(struct_offset / sizeof(*this))).c_str());
-        }
+        auto name = std::string("meter #") + std::to_string(struct_offset / sizeof(*this));
+        CHECK_INTERFACE_BITMAP_SEQ(this->meter_bitmap, this->sequence_index, name.c_str());
+        CHECK_HIGH_RES_SCALING(this->scaling_flags, name.c_str());
     }
 
     void WeaponHUDInterfaceStaticElement::post_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t struct_offset) {
         // Make sure it's valid
-        if(this->sequence_index != NULL_INDEX && (!workload.disable_recursion || !workload.disable_error_checking)) {
-            CHECK_INTERFACE_BITMAP_SEQ(this->interface_bitmap, this->sequence_index, (std::string("static element #") + std::to_string(struct_offset / sizeof(*this))).c_str());
-        }
+        auto name = std::string("static element #") + std::to_string(struct_offset / sizeof(*this));
+        CHECK_INTERFACE_BITMAP_SEQ(this->interface_bitmap, this->sequence_index, name.c_str());
+        CHECK_HIGH_RES_SCALING(this->scaling_flags, name.c_str());
     }
 
     void WeaponHUDInterfaceOverlayElement::post_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t, std::size_t struct_offset) {
         // Make sure it's valid
         std::size_t overlay_count = this->overlays.size();
         for(std::size_t i = 0; i < overlay_count; i++) {
+            auto name = std::string("overlay #") + std::to_string(struct_offset / sizeof(*this));
             auto &overlay = this->overlays[i];
-            CHECK_INTERFACE_BITMAP_SEQ(this->overlay_bitmap, overlay.sequence_index, (std::string("overlay #") + std::to_string(struct_offset / sizeof(*this))).c_str());
+            CHECK_INTERFACE_BITMAP_SEQ(this->overlay_bitmap, overlay.sequence_index, name.c_str());
+            CHECK_HIGH_RES_SCALING(overlay.scaling_flags, name.c_str());
         }
     }
     
@@ -168,6 +177,15 @@ namespace Invader::Parser {
         CHECK_INTERFACE_BITMAP_SEQ(this->motion_sensor_background_interface_bitmap, this->motion_sensor_background_sequence_index, "motion sensor background");
         CHECK_INTERFACE_BITMAP_SEQ(this->motion_sensor_foreground_interface_bitmap, this->motion_sensor_foreground_sequence_index, "motion sensor foreground");
         
+        // Check these memes
+        CHECK_HIGH_RES_SCALING(this->hud_background_scaling_flags, "HUD background");
+        CHECK_HIGH_RES_SCALING(this->shield_panel_background_scaling_flags, "shield panel background");
+        CHECK_HIGH_RES_SCALING(this->shield_panel_meter_scaling_flags, "shield panel meter");
+        CHECK_HIGH_RES_SCALING(this->health_panel_background_scaling_flags, "health panel background");
+        CHECK_HIGH_RES_SCALING(this->health_panel_meter_scaling_flags, "health panel meter");
+        CHECK_HIGH_RES_SCALING(this->motion_sensor_background_scaling_flags, "motion sensor background");
+        CHECK_HIGH_RES_SCALING(this->motion_sensor_foreground_scaling_flags, "motion sensor foreground");
+        
         // Bounds check the overlays
         auto overlay_count = this->overlays.size();
         for(std::size_t i = 0; i < overlay_count; i++) {
@@ -175,6 +193,7 @@ namespace Invader::Parser {
             std::snprintf(overlay_name, sizeof(overlay_name), "overlay #%zu", i);
             auto &overlay = this->overlays[i];
             CHECK_INTERFACE_BITMAP_SEQ(overlay.interface_bitmap, overlay.sequence_index, overlay_name);
+            CHECK_HIGH_RES_SCALING(overlay.scaling_flags, overlay_name);
         }
         
         // Bounds check the meters
@@ -182,9 +201,14 @@ namespace Invader::Parser {
         for(std::size_t i = 0; i < meter_count; i++) {
             char meter_name[256];
             std::snprintf(meter_name, sizeof(meter_name), "meter #%zu", i);
+            auto meter_bg_name = std::string(meter_name) + " background";
+            auto meter_fg_name = std::string(meter_name) + " meter";
+            
             auto &meter = this->meters[i];
-            CHECK_INTERFACE_BITMAP_SEQ(meter.background_interface_bitmap, meter.background_sequence_index, (std::string(meter_name) + " background").c_str());
-            CHECK_INTERFACE_BITMAP_SEQ(meter.meter_meter_bitmap, meter.meter_sequence_index, (std::string(meter_name) + " meter").c_str());
+            CHECK_INTERFACE_BITMAP_SEQ(meter.background_interface_bitmap, meter.background_sequence_index, meter_bg_name.c_str());
+            CHECK_INTERFACE_BITMAP_SEQ(meter.meter_meter_bitmap, meter.meter_sequence_index, meter_fg_name.c_str());
+            CHECK_HIGH_RES_SCALING(meter.background_scaling_flags, meter_bg_name.c_str());
+            CHECK_HIGH_RES_SCALING(meter.meter_scaling_flags, meter_fg_name.c_str());
         }
     }
 }